Updated LimitDefault and LimitMax type from uint32 to uint64

+++ common/txmgr/types/config.go
@@ -42,7 +42,7 @@ type BroadcasterListenerConfig interface {
 
 type ConfirmerFeeConfig interface {
 	BumpTxDepth() uint32
-	LimitDefault() uint32
+	LimitDefault() uint64
 
 	// from gas.Config
 	BumpThreshold() uint64
