+++ core/services/vrf/v2/bhs_feeder_test.go
@@ -57,7 +57,7 @@ func TestStartHeartbeats(t *testing.T) {
 		c.EVM[0].MinIncomingConfirmations = ptr[uint32](2)
 		c.Feature.LogPoller = ptr(true)
 		c.EVM[0].FinalityDepth = ptr[uint32](2)
-		c.EVM[0].GasEstimator.LimitDefault = ptr(uint32(gasLimit))
+		c.EVM[0].GasEstimator.LimitDefault = ptr(uint64(gasLimit))
 		c.EVM[0].LogPollInterval = commonconfig.MustNewDuration(time.Second)
 	})
